move json parsing implementation outside of header files
