+++ test/unit_test/test_common.h
@@ -8,6 +8,7 @@
 #include "vw/core/ccb_reduction_features.h"
 #include "vw/core/parse_example_json.h"
 #include "vw/core/slates_label.h"
+#include "vw/core/shared_data.h"
 #include "vw/core/vw.h"
 
 #include <boost/test/test_tools.hpp>
