+++ vowpalwabbit/core/include/vw/core/cache.h
@@ -6,8 +6,6 @@
 
 #include "vw/core/vw_fwd.h"
 
-#include <cstdint>
-
 namespace VW
 {
 namespace details
